refactor: isolate URLProtocolStub in its own file for better organization
