+++ Modules/Networking/NetworkingTests/Helpers/URLProtocolStub.swift
@@ -1,0 +1,75 @@
+//
+//  URLProtocolStub.swift
+//  NetworkingTests
+//
+//  Created by MacPro on 28/06/25.
+//  Copyright © 2025 Cesar Lima Consulting. All rights reserved.
+//
+
+import Foundation
+
+final class URLProtocolStub: URLProtocol {
+    private static var stub: Stub?
+    private static var requestObserver: ((URLRequest) -> Void)?
+    private(set) static var requestsCount: Int = 0
+    
+    private struct Stub {
+        let data: Data?
+        let response: URLResponse?
+        let error: Error?
+    }
+    
+    static func stub(data: Data?, response: URLResponse?, error: Error?) {
+        stub = Stub(data: data, response: response, error: error)
+    }
+    
+    static func observeRequest(observer: @escaping (URLRequest) -> Void) {
+        requestObserver = observer
+    }
+    
+    static func startInterceptingRequests() {
+        URLProtocol.registerClass(URLProtocolStub.self)
+    }
+    
+    static func stopInterceptingRequests() {
+        URLProtocol.unregisterClass(URLProtocolStub.self)
+        stub = nil
+        requestObserver = nil
+        requestsCount = 0
+    }
+    
+    override class func canInit(with request: URLRequest) -> Bool {
+        return true
+    }
+    
+    override class func canonicalRequest(for request: URLRequest) -> URLRequest {
+        request
+    }
+    
+    override func startLoading() {
+        URLProtocolStub.requestsCount += 1
+        
+        if let requestObserver = URLProtocolStub.requestObserver {
+            requestObserver(request)
+        }
+        
+        if let stub = URLProtocolStub.stub {
+            if let response = stub.response {
+                client?.urlProtocol(self, didReceive: response, cacheStoragePolicy: .notAllowed)
+            }
+            
+            if let data = stub.data {
+                client?.urlProtocol(self, didLoad: data)
+            }
+            
+            if let error = stub.error {
+                client?.urlProtocol(self, didFailWithError: error)
+                return
+            }
+        }
+        
+        client?.urlProtocolDidFinishLoading(self)
+    }
+    
+    override func stopLoading() {}
+}
